Clean-up endpoints.
Move existing endpoints to separate class.

Work on #3
